src/main/java/leetcode/math/problem202/HappyNumber.java
package leetcode.math.problem202;

import java.util.HashSet;
import java.util.Set;

public class HappyNumber {

    public boolean isHappy(int n) {
        Set<Integer> set = new HashSet<>();
        int sum;
        while (true) {
            sum = doSum(n);
            if (sum == 1) {
                return true;
            } else if (set.contains(sum)) {
                return false;
            } else {
                n = sum;
                set.add(sum);
            }
        }
    }

    private int doSum(int n) {
        int sum = 0;
        while (n > 0) {
            int remainder = n % 10;
            sum += remainder * remainder;
            n /= 10;
        }
        return sum;
    }
}
